+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# 0.0.3
+
+Features:
+
+- Added two strategies to get best representative features of each cluster:
+    - Strategy based on mutual information with target.
+    - Strategy based on Forward Feature Selection using clusters and RandomForest.
+
+
 # 0.0.2
 
 Features: